requirements.txt
python-dotenv
psycopg2-binary
flask
flask_sqlalchemy
requests
sqlalchemy
aws_secretsmanager_caching
botocore
sqlalchemy_utils
flask-cors